stackexchange/requirements.txt
azure-functions
python-dotenv==0.13.0
requests==2.23.0
grpcio==1.27.2